requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.11.11
qrcode[pil]==8.0
